+++ CHANGELOG.md
@@ -6,6 +6,14 @@ The format is based on [Keep a
 Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
+### Fixed
+
+- Fixed some stray references to structs which should be typed as records
+- Fixed `Spear.Event.to_checkpoint/1` to carry over the `:subscription` key
+  from a `t:Spear.Event.t/0`'s metadata
+
 ## 0.8.0 - 2021-04-26
 
 ### Added